Reimplement channel groupings as sharing of a structure ...
... among a contiguous sub-sequence of the tracks in a list, not a boolean
flag on the first member of the group.

The group data structure is empty for now, simply using its memory address as
an identifier for a group of tracks.  Future commits will move some fields
out of the tracks and into the group data.  Such values, like the name, are not
intended to differ among channels of a group.